Integrate register form into auth modal
